+++ src/openrct2/TrackImporter.cpp
@@ -35,4 +35,4 @@ namespace TrackImporter
     {
         return String::Equals(extension, ".td4", true);
     }
-} // namespace ParkImporter
+} // namespace TrackImporter
